+++ instrumentation/spring/spring-data/spring-data-1.8/javaagent/build.gradle.kts
@@ -29,8 +29,6 @@ dependencies {
   compileOnly(project(":instrumentation-annotations-support"))
 
   testInstrumentation(project(":instrumentation:jdbc:javaagent"))
-  testInstrumentation(project(":instrumentation:r2dbc-1.0:javaagent"))
-  testInstrumentation(project(":instrumentation:reactor:reactor-3.1:javaagent"))
 
   testImplementation(project(":instrumentation:spring:spring-data:spring-data-common:testing"))
 
@@ -47,27 +45,10 @@ dependencies {
   latestDepTestLibrary("org.springframework:spring-test:5.+")
 }
 
-testing {
-  suites {
-    val reactiveTest by registering(JvmTestSuite::class) {
-      dependencies {
-        implementation("org.springframework.data:spring-data-r2dbc:3.0.0")
-        implementation("org.testcontainers:testcontainers")
-        implementation("io.r2dbc:r2dbc-h2:1.0.0.RELEASE")
-        implementation("com.h2database:h2:1.4.197")
-      }
-    }
-  }
-}
-
 tasks {
   test {
     jvmArgs("--add-opens=java.base/java.lang.invoke=ALL-UNNAMED")
     jvmArgs("--add-opens=java.base/java.lang=ALL-UNNAMED")
     jvmArgs("-XX:+IgnoreUnrecognizedVMOptions")
-  }
-
-  check {
-    dependsOn(testing.suites)
   }
 }
