+++ instrumentation/spring/spring-data/spring-data-3.0/testing/src/reactiveTest/java/io/opentelemetry/javaagent/instrumentation/spring/data/v3_0/repository/Customer.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.spring.data.v1_8.repository;
+package io.opentelemetry.javaagent.instrumentation.spring.data.v3_0.repository;
 
 import java.util.Objects;
 import javax.annotation.Nullable;
